+++ Class/Packet/n_frame.cpp
@@ -4,18 +4,18 @@
 n_Frame::n_Frame() { this->data = nullptr; }
 
 // constructor with uint8_t*
-n_Frame::n_Frame(uint8_t* data, size_t len) {
+n_Frame::n_Frame(uint8_t* data, int len) {
     this->data = new uint8_t [len];
     if (this->data != nullptr)
-        memcpy(this->data, data, len);
+        memcpy(this->data, data, static_cast<size_t>(len));
     this->length = len;
 }
 
 // constructor with const uint8_t*
-n_Frame::n_Frame(const uint8_t* data, size_t len) {
+n_Frame::n_Frame(const uint8_t* data, int len) {
     this->data = new uint8_t [len];
     if (this->data != nullptr)
-        memcpy(this->data, data, len);
+        memcpy(this->data, data, static_cast<size_t>(len));
     this->length = len;
 }
 
@@ -26,31 +26,31 @@ n_Frame::~n_Frame(){
 }
 
 // set length of packet
-void n_Frame::setLength(size_t len){
+void n_Frame::setLength(int len){
     this->length = len;
 }
 
 // return length of packet
-size_t n_Frame::getLength() const {
+int n_Frame::getLength() const {
     return this->length;
 }
 
 // set data with uint8_t*
-void n_Frame::setFrameData(uint8_t* data, size_t len){
+void n_Frame::setFrameData(uint8_t* data, int len){
     if (this->data != nullptr){
         delete[] this->data;
     }
     this->data = new uint8_t [len];
-    memcpy(this->data, data, len);
+    memcpy(this->data, data, static_cast<size_t>(len));
 }
 
 // set data with const uint8_t*
-void n_Frame::setFrameData(const uint8_t* data, size_t len){
+void n_Frame::setFrameData(const uint8_t* data, int len){
     if (this->data != nullptr){
         delete[] this->data;
     }
     this->data = new uint8_t [len];
-    memcpy(this->data, data, len);
+    memcpy(this->data, data, static_cast<size_t>(len));
 }
 
 // return data
