+++ Class/Packet/n_frame.h
@@ -8,18 +8,18 @@ class n_Frame
 {
 public:
     n_Frame();
-    n_Frame(uint8_t* data, size_t len);
-    n_Frame(const uint8_t* data, size_t len);
+    n_Frame(uint8_t* data, int len);
+    n_Frame(const uint8_t* data, int len);
     virtual ~n_Frame();
-    void setLength(size_t len);
-    size_t getLength() const ;
-    void setFrameData(uint8_t* data, size_t len);
-    void setFrameData(const uint8_t* data, size_t len);
+    void setLength(int len);
+    int getLength() const ;
+    void setFrameData(uint8_t* data, int len);
+    void setFrameData(const uint8_t* data, int len);
     uint8_t* getFrameData() const ;
     virtual std::string what() const { return "Frame"; } //return name of class
 
 private:
-    size_t length;
+    int length;
     uint8_t* data;
 };
